-Add setters and getters for Inventory Manager and PlotManager -Add rent attribute to InventoryManager -Add increaseRent and removeRent method to InventoryManager and add a call of increaseRent in PlotManager.createNewPlot. -State class implementation
